requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
flask
fastapi
statsmodels
ipykernel
PyYAML
dill
shap
jinja2 
python-multipart
uvicorn
xgboost
mlflow
xlsxwriter
python-dotenv
imblearn
#-e .